Refactor agents and fixtures for improved test automation
- Renamed 'api-test-automation-openapi' to 'api-test-automation' with updated description.
- Deleted obsolete 'playwright-expert-detailed' agent.
- Removed 'test-automation-architect' agent to streamline roles.
- Updated 'test-planner' agent description for clarity.
- Added new 'accessibility-expert' agent focused on WCAG compliance.
- Introduced 'ui-test-automation-expert' agent for Playwright UI automation.
- Refactored fixtures to 'index.fixture.ts' for better organization and functionality.
- Updated test files to utilize new fixture structure.
- Enhanced .gitignore to include temporary AI files.
