+++ .gitignore
@@ -8,3 +8,4 @@ node_modules/
 /playwright/.auth/
 .playwright-mcp/
 *.png
+.ai-temp/
